Player Engine Movement
Player engine movement v1.0

You can use the L/R bumpers to turn, or hold down both to go forward.
This is really early stuff, so a lot of work needs to be done.

A few things were added to BaseShip.cs, mainly engine things. In the
player script, it actually fires the engines (along with a few hidden
ones), and locks the ship to the camera view, just so you don't fly off
screen.

Added Engine scripts, to turn on/off and apply forces.

The turrets decidedly do not like the ships spinning, so that should
probably be looked at next