Auto turrets much more accurate
Enemies will be next I suppose.
On a side note, the original implementations made auto turrets super
super accurate, so I've added a little randomness in, so it isn't as
painstakingly accurate.
The number (the two random functions in turret fire) can be fine tuned
if we want more or less accurate turrets